Update projet and get every API objects by Injection
